Break out Utils.scala into smaller portions
